requirements.txt
flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
requests>=2.28.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
groq>=0.5.0